reorganized API code
