+++ lib/constants.js
@@ -1,10 +1,6 @@
-import { isDev } from "./utils.js";
+import dotenv from "dotenv";
 
-if (isDev()) {
-  const dotenv = await import("dotenv");
-
-  dotenv.config();
-}
+dotenv.config();
 
 export const API_TOKEN = process.env.CLOUDFLARE_API_KEY;
